KL-147/refactor: separate fetching logic from TagSwitchList
- ProductList의 파일 구조와 유사하게 변경하였습니다.
- TagSwitchList로부터 fetching 작업을 분리하였습니다.
	- 분리된 fetching 작업이 새로운 TagDataFetcher가 되었습니다.
	- TagDataFetcher의 기존 동작은 훅으로 분리하였습니다.
